pyproject.toml
[tool.poetry]
name = "netrise-provenance-sdk"
version = "0.1.1"
description = "Provenance API Python SDK for package, repository, and contributor intelligence"
authors = ["NetRise <support@netrise.io>"]
readme = "README.md"
packages = [
  { include = "netrise_provenance_sdk", from = "src" },
  { include = "netrise_provenance_sdk_client", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.12"
httpx = ">=0.27.0,<1.0.0"
pydantic = ">=2.0.0,<3.0.0"
python-dotenv = ">=1.0.0,<2.0.0"

[tool.poetry.group.dev.dependencies]
openapi-python-client = {version = ">=0.21.0,<1.0.0", python = "<4"}
pytest = ">=8.0.0,<9.0.0"
pytest-cov = "^7.0.0"
respx = ">=0.21.0,<1.0.0"
twine = "^6.2.0"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
